chore(core): migrate eligible jvmTest to commonTest (#401) (#443)
* chore(core): migrate 5 util jvmTest files to commonTest (#401)

Migrate Array2DMapTest, DoubletonTest, EnumUnorientedGraphTest,
HashMapGraphTest, and MultimapExtensionsTest from jvmTest to commonTest
for native (linuxX64) test coverage.

Changes:
- Replace java.util.* imports with Kotlin stdlib equivalents
- Replace JUnit5 annotations with kotlin.test (@BeforeTest, @Test)
- Flatten @Nested/@DisplayName structure to plain test methods
- Replace java.util.Random with kotlin.random.Random
- Replace Integer.signum with Int.sign
- Replace TreeMap/sortedMapOf (JVM-only) with mutableMapOf
- Remove dead performance benchmark code (tst/avg methods)
- Add shared assertion extensions to core-test CommonAssertKExtensions:
  hasMessageContaining, containsEntry, doesNotContainValue, allMatch,
  containsNull, and block assertThat

All 1712 jvmTests, 493 linuxX64Tests, and 156 integration tests pass.


* fix(core): address PR #443 review findings

- CommonAssertKExtensions: keep @JvmName (needed for JVM erasure), add docs
- Array2DMapTest: document mutableMapOf semantics vs TreeMap
- MultimapExtensionsTest: containsExactlyInAnyOrder -> containsExactly for order test
- DoubletonTest: use @BeforeTest instead of manual setup calls
- EnumUnorientedGraphTest: fix mismatched test name, remove redundant casts


* fix(core): address PR #443 round 2 review findings (#401)

- @JvmName: kept annotation (removal causes JVM signature clash) but
  shortened the explanatory comment to focus on the actual erasure reason
- keys_sortedByNaturalOrder: renamed to keys_afterOutOfOrderInsertion_containsAllKeys
  and removed misleading sorted map comments
- treeMap variable: renamed to referenceMap throughout Array2DMapTest


---------
